+++ lib/olels/pages/instance_editor.rb
@@ -19,7 +19,7 @@ module OLE_QA::Framework::OLELS
     def set_elements
       super
       # Override :message element on Editor base class.
-      element(:message, force = true)                 {b.span(:id => "workInstanceMessage_span")}
+      element(:message)                               {b.span(:id => "workInstanceMessage_span")}
       element(:location_toggle)                       {b.a(:id => "OleLocationInformationSection_toggle")}
       element(:location_field)                        {b.text_field(:id => "OleHoldingLocation_control")}
       element(:call_number_toggle)                    {b.a(:id => "OleCallNumberInformation_toggle")}