+++ lib/olels/pages/item_editor.rb
@@ -19,7 +19,7 @@ module OLE_QA::Framework::OLELS
     def set_elements
       super
       # Override message in Editor base class, add multi-message element.
-      element(:message, true)                             {b.span(:id => "workItemMessage_span")}
+      element(:message)                                   {b.span(:id => "workItemMessage_span")}
       element(:messages)                                  {b.span(:id => "workItemMessage_span").font}
       # Holdings Location/Call Number Information
       element(:holdings_location_field)                   {b.span(:id => "OleHoldingLocationLevelName_control")}